Fix larastan on level 3 (#561)
* Fix variable declarations

* Fix larastan on level 2

* Fix code formatting issues

* Fix token issues when seeding tenants

* Fix token issues when seeding tenants

* Fix PHPCS issue

* Remove newly created relations

* Update expiration time

* Set larastan to level 3

* Fix larastan on level 3

* Fix larastan on level 3

* Reset larastan level

* Fix code formatting

* Fix type hinting

+++ src/backend/app/modules/Transaction/Provider/AgentTransactionProvider.php
@@ -36,7 +36,7 @@ class AgentTransactionProvider implements ITransactionProvider {
         $this->agentTransaction->device_id = (int) $data['device_id'];
 
         // common transaction data
-        $this->transaction->amount = (float) $data['amount'];
+        $this->transaction->amount = (int) $data['amount'];
         $this->transaction->sender = 'Agent-'.$data['agent_id'];
         $this->transaction->message = $data['device_serial'];
         $this->transaction->type = 'energy';
